// src/utils/enums.ts
export enum SocialsEnum {
    FB = 'fb',
    X = 'x',
    YT = 'yt',
}

export enum AppButtonColorEnum {
    WHITE = 'white',
    ACCENT = 'accent',
}

export enum PostModesEnum {
    HOME = 'home',
    POST = 'post',
}

export enum FormFieldsEnum {
    EMAIL = 'email',
    PASSWORD = 'password',
    FIRST_NAME = 'first_name',
    LAST_NAME = 'last_name',
}

export enum FormTypesEnum {
    SIGN_IN = 'sign-in',
    SIGN_UP = 'sign-up',
}